third_party/blink/web_tests/external/wpt/mediacapture-insertable-streams/idlharness.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] MediaStreamTrackProcessor interface: existence and properties of interface object
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface object length
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface object name
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor interface: attribute readable
  assert_own_property: self does not have own property "MediaStreamTrackProcessor" expected property "MediaStreamTrackProcessor" missing
[FAIL] MediaStreamTrackProcessor must be primary interface of new MediaStreamTrackProcessor({ track: new VideoTrackGenerator() })
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: MediaStreamTrackProcessor is not defined"
[FAIL] Stringification of new MediaStreamTrackProcessor({ track: new VideoTrackGenerator() })
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: MediaStreamTrackProcessor is not defined"
[FAIL] MediaStreamTrackProcessor interface: new MediaStreamTrackProcessor({ track: new VideoTrackGenerator() }) must inherit property "readable" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: MediaStreamTrackProcessor is not defined"
[FAIL] VideoTrackGenerator interface: existence and properties of interface object
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface object length
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface object name
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: attribute writable
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: attribute muted
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator interface: attribute track
  assert_own_property: self does not have own property "VideoTrackGenerator" expected property "VideoTrackGenerator" missing
[FAIL] VideoTrackGenerator must be primary interface of new VideoTrackGenerator()
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VideoTrackGenerator is not defined"
[FAIL] Stringification of new VideoTrackGenerator()
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VideoTrackGenerator is not defined"
[FAIL] VideoTrackGenerator interface: new VideoTrackGenerator() must inherit property "writable" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VideoTrackGenerator is not defined"
[FAIL] VideoTrackGenerator interface: new VideoTrackGenerator() must inherit property "muted" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VideoTrackGenerator is not defined"
[FAIL] VideoTrackGenerator interface: new VideoTrackGenerator() must inherit property "track" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VideoTrackGenerator is not defined"
Harness: the test ran to completion.
